add tester

+++ ft_puthex.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_puthex.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ncontin <ncontin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: noscrub <noscrub@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 08:56:56 by ncontin           #+#    #+#             */
-/*   Updated: 2024/10/23 12:58:43 by ncontin          ###   ########.fr       */
+/*   Updated: 2024/10/23 16:31:11 by noscrub          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex(int n, char hex_case)
+int	ft_puthex(unsigned long int n, char hex_case)
 {
 	int	counter;
 	int	rem;
